utils/icd10/requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.2
PyYAML==6.0
soupsieve==2.4.1
